rename resolver to composer
