+++ lib/composer.js
