+++ lib/composer.test.js
