+++ lib/html-composer.js
@@ -14,7 +14,7 @@ const _ = require('lodash'),
   log = require('./log'),
   multiplexTemplates = require('multiplex-templates'),
   references = require('./services/references'),
-  resolver = require('./resolver'),
+  composer = require('./composer'),
   media = require('./media');
 
 /**
@@ -136,7 +136,7 @@ function renderTemplate() {
  * @returns {Promise}
  */
 function renderByConfiguration(options, res) {
-  return resolver.resolveComponentReferences(options.data)
+  return composer.resolveComponentReferences(options.data)
     .then(applyOptions(options, res))
     .then(renderTemplate());
 }
