+++ lib/services/pages.js
@@ -5,7 +5,7 @@ const _ = require('lodash'),
   db = require('../services/db'),
   bluebird = require('bluebird'),
   references = require('../services/references'),
-  resolver = require('../resolver'),
+  composer = require('../composer'),
   siteService = require('./sites'),
   notifications = require('./notifications'),
   uid = require('../uid');
@@ -58,7 +58,7 @@ function getPageClonePutOperations(pageData, locals) {
       //for all strings that are component references
       promises.push(components.get(pageValue, locals).then(function (refData) {
         //only follow the paths of instance references.  Don't clone default components
-        return resolver.resolveComponentReferences(refData, isInstanceReferenceObject);
+        return composer.resolveComponentReferences(refData, isInstanceReferenceObject);
       }).then(function (resolvedData) {
         //for each instance reference within resolved data
         _.each(_.listDeepObjects(resolvedData, isInstanceReferenceObject), function (obj) {
@@ -143,7 +143,7 @@ function getRecursivePublishedPutOperations(locals) {
      * 4) Get list of put operations needed
      */
     return components.get(rootComponentRef, locals)
-      .then(resolver.resolveComponentReferences)
+      .then(composer.resolveComponentReferences)
       .then(references.replaceAllVersions(published))
       .then(function (data) {
         return components.getPutOperations(references.replaceVersion(rootComponentRef, published), data, locals);
